refactor: alterando valor para BigDecimal alinhando com a regra definida em conta-service
Erro bobo por falta de atenção, usar double ou float pode causar erros de aproximação e casas decimais, por isso BigDecimal entrou em cena novamente para comunicar-se com conta-service.

+++ custodia-service/src/main/java/com/example/custodia_service/domain/AtivoCustodia.java
@@ -4,6 +4,7 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Entity
@@ -14,13 +15,13 @@ public class AtivoCustodia {
     private UUID id;
 
     private String cpfCliente;
-    private Double valor;
+    private BigDecimal valor;
     private String statusCustodia;
 
     public AtivoCustodia() {
     }
 
-    public AtivoCustodia(String cpfCliente, Double valor, String statusCustodia) {
+    public AtivoCustodia(String cpfCliente, BigDecimal valor, String statusCustodia) {
         this.cpfCliente = cpfCliente;
         this.valor = valor;
         this.statusCustodia = statusCustodia;
@@ -42,11 +43,11 @@ public class AtivoCustodia {
         this.cpfCliente = cpfCliente;
     }
 
-    public Double getValor() {
+    public BigDecimal getValor() {
         return valor;
     }
 
-    public void setValor(Double valor) {
+    public void setValor(BigDecimal valor) {
         this.valor = valor;
     }
